requirements.txt
# Core Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Database
pg8000==1.30.3

# Authentication
PyJWT==2.8.0

# External APIs
openai==1.3.7
requests==2.31.0

# Environment Management
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0